+++ lib/madmin/fields/password.rb
@@ -1,0 +1,6 @@
+module Madmin
+  module Fields
+    class Password < Field
+    end
+  end
+end