+++ test/dummy/app/models/user.rb
@@ -1,5 +1,7 @@
 class User < ApplicationRecord
   has_many :posts
+  accepts_nested_attributes_for :posts, allow_destroy: true
+
   has_many :comments
   has_and_belongs_to_many :habtms, join_table: :user_habtms
 
@@ -8,4 +10,7 @@ class User < ApplicationRecord
   attribute :virtual_attribute, default: "virtual"
 
   has_one_attached :avatar
+
+  has_secure_password
+  has_secure_token
 end