+++ test/dummy/db/migrate/20200911011935_create_users.rb
@@ -4,6 +4,8 @@ class CreateUsers < ActiveRecord::Migration[6.0]
       t.string :first_name
       t.string :last_name
       t.date :birthday
+      t.string :password_digest
+      t.string :token
 
       t.timestamps
     end